# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/business/{id}": {
      "get": {
        "description": "Retrieves a business by its unique identifier",
        "operationId": "BusinessController_getBusinessById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Business ID",
            "schema": {
              "example": "business-123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Business found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            }
          },
          "404": {
            "description": "Business not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get business by ID",
        "tags": [
          "business"
        ]
      },
      "patch": {
        "description": "Updates an existing business",
        "operationId": "BusinessController_updateBusiness",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Business ID",
            "schema": {
              "example": "business-123"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Business updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            }
          },
          "404": {
            "description": "Business not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update business",
        "tags": [
          "business"
        ]
      }
    },
    "/business/keyName/{keyName}": {
      "get": {
        "description": "Retrieves a business by its unique key name",
        "operationId": "BusinessController_getBusinessByKeyName",
        "parameters": [
          {
            "name": "keyName",
            "required": true,
            "in": "path",
            "description": "Business key name",
            "schema": {
              "example": "my-business"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Business found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessKeyNameDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get business by key name",
        "tags": [
          "business"
        ]
      }
    },
    "/business": {
      "get": {
        "description": "Retrieves a list of all businesses",
        "operationId": "BusinessController_getBusiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of businesses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Business"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all businesses",
        "tags": [
          "business"
        ]
      },
      "post": {
        "description": "Creates a new business account",
        "operationId": "BusinessController_createBusiness",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Business created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new business",
        "tags": [
          "business"
        ]
      }
    },
    "/business/{id}/whatsapp-connection": {
      "get": {
        "operationId": "BusinessController_getWhatsappConnectionById",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business"
        ]
      },
      "patch": {
        "operationId": "BusinessController_updateWhatsappConnection",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsappConnection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business"
        ]
      }
    },
...
    "/rol": {
      "get": {
        "description": "Retrieves a list of all roles",
        "operationId": "RolController_getRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rol"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all roles",
        "tags": [
          "rol"
        ]
      },
      "post": {
        "description": "Creates a new role with specified permissions",
        "operationId": "RolController_createRol",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "manager"
                  },
                  "description": {
                    "type": "string",
                    "example": "Manager role"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rol"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new role",
        "tags": [
          "rol"
        ]
      }
    },
    "/rol/{id}": {
      "get": {
        "description": "Retrieves a role by its unique identifier",
        "operationId": "RolController_getRolById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "example": "rol-123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rol"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get role by ID",
        "tags": [
          "rol"
        ]
      }
    },
...
    "/patient-history/control/{id}": {
      "patch": {
        "description": "Updates control information for a patient history",
        "operationId": "PatientHistoryController_updatePatientHistoryControl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Patient history ID",
            "schema": {
              "example": "patient-history-123"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientHistory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient history control updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientHistory"
                }
              }
            }
          },
          "404": {
            "description": "Patient history not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update patient history control",
        "tags": [
          "patient-history"
        ]
      }
    },
...
  },
  "info": {
    "title": "ESV Backend API",
    "description": "API documentation for ESV Backend - Appointment and Queue Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://app.easuavez.com",
      "description": "Default server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "Business": {
        "type": "object",
        "properties": {}
      },
      "BusinessKeyNameDetailsDto": {
        "type": "object",
        "properties": {}
      },
      "WhatsappConnection": {
        "type": "object",
        "properties": {}
      },
...
      "Rol": {
        "type": "object",
        "properties": {}
      },
...
      "PatientHistory": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}